Added brand user skeleton and apparel form wizard

+++ app/src/main/java/com/stephben/hypewear/apparel/data/dtos/EcoMetricsDto.kt
@@ -2,7 +2,7 @@ package com.stephben.hypewear.apparel.data.dtos
 
 data class EcoMetricsDto(
     val materialSustainability: Double? = null,
-    val carbonFootprint: Int? = null,
-    val waterFootprint: Int? = null,
+    val carbonFootprint: Double? = null,
+    val waterFootprint: Double? = null,
     val packagingSustainability: Double? = null,
 )
